getAllUsers working with Firebase
